+++ app/src/main/java/app/revanced/manager/util/FileFilters.kt
@@ -13,3 +13,8 @@ fun isAllowedPatchBundleFile(path: Path): Boolean {
     val extension = path.name.substringAfterLast('.', "").lowercase(Locale.ROOT)
     return extension == "rvp" || extension == "mpp" || extension == "arp"
 }
+
+fun isAllowedSplitArchiveFile(path: Path): Boolean {
+    val extension = path.name.substringAfterLast('.', "").lowercase(Locale.ROOT)
+    return extension in SPLIT_ARCHIVE_FILE_EXTENSIONS
+}